fix(backend/cryptomator): Make RemoveDir return err for Bucketbased
BucketBased remotes just return nil when trying to delete a
subdirectory. Due to the directory structure of cryptomator
requiring directory id files this is suboptimal.

When we try to delete a directory on BucketBased remotes no
error is returned which then leads to cryptomator deleting the
directory id file resulting in inaccessable data.

We handle this separately now and try to list the files first
and check whether the directory is empty.